pyproject.toml
[tool.poetry]
name = "keycloak"
version = "0.1.0"
description = ""
authors = ["Christian Lyons <christian.lyons@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.4"
requests = "^2.32.3"
python-keycloak = "^5.3.1"
pytest-dotenv = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
